pyproject.toml
[project]
name = "cert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-docx>=0.8.11",
]
